Finished functionality for prefixer
Ran tests

Merged latest develop
